Replace country with site grain
To allow for multiple sites per country (for example in case of
locations in multiple cities in one country, or even multiple
locations in one city), and to make mapping of locations to the
DCIM/IPAM data (which already uses the site codes) easier, replace
the "country" grain with a "site" one and refactor logic related to
it to reflect the change.

References are replaced with the respective, already defined, site
codes.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>