Certificates for MariaDB servers
To facilitate TLS based MySQL, add internal certificates for all
Galera nodes - mybackup does not usually receive client access, but
was added as well, as it does not hurt to have it covered too.
To allow for certificate renewals without downtime, add logic to
send the respective TLS flush command instead of a service restart
for hosts with "mariadb" certificate service entries.
A user with the relevant privileges matching the certificate
deployment user is added to allow for rootless execution.

Note this only covers the backend servers - TLS for the Galera
frontend can be added later, as it may require adjustment of the
proxy configuration.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>