Configure BuddyCheck to protect os-update
This enables a service drop-in for os-update on nodes part of HA pairs
having the "profile:buddycheck:buddy" pillar set, making the os-update
service first call buddycheck to assess wehther the HA partner node is
healthy. If this assessment is negative or fails, os-update will not
continue to install updates. This protects us from bad updates - if a
HA node shows failing services after installing updates, the other node
will not proceed to install the same, potentially dangerous, updates,
allowing an administrator to inspect what happened to the first node
first.

https://progress.opensuse.org/issues/165935

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>