Consider drop-ins in daemon reload check
Only searching for *.service is not sufficient in case of services where
only the unit dropin and not the service unit itself is managed by Salt -
if Salt modifies a foo.service.d/bar.conf drop-in file, foo.service must
be considered in the NeedDaemonReload query in order for the changes to
always be fully applied by triggering a systemd daemon reload.

In cases where both the main unit as well as a drop-in are managed by
Salt, the query will receive the same specifier twice - this is not an
issue, as it will just yield additional yes/no lines in the pipe for
grep.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>