Increase DehydratedHookLastRun alert days
Solves this scenario:
Dehydrated hook runs on Sep 6 at 00:38:08 GMT.
But skips renew due to cert being valid till Oct 6 00:52:34 2024 GMT.
(Renewal occurs 30 days prior to expiry)

On Sep 7, the dehydrated timer did not run because the machine rebooted for updates.
This suspicion is further reinforced by the fact that it didn't run on Sep 5 as well.
There was also a reboot due to updates on Sep 5.

Increasing the DehydratedHookLastRun from 60 to 63 days should account for any such situations.

Signed-off-by: Pratyush Desai <pratyush.desai@liberta.casa>