Import SSH keys for Galera backup initialization
The backup_replica_sync.sh script on galera1 connects to mybackup to
bootstrap the MariaDB backup replica. This requires it to copy files and
execute commands as root on the target system.
It could be improved using a less privileged user and sudo rules, but
would require refactoring of the script and additional testing.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>