Use PROXY for factory-dashboard rsyncd
Enable the use of the PROXY protocol to show the real client IP
addresses on the backend, which are more useful for inspection of logs
and allow for more granular and hence more secure ACL filtering than
having all traffic covered by the proxy addresses.
This aligns the setup with the manpages rsyncd and should then be
the standard for all rsyncd's hosted behind HAProxy.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>