Allow SSH from galera1 to mybackup as root
For the backup initialization script, the systems connect between each
other as root. The keys for this were imported, but the matching
override for our default sshd configuration was missing.
To improve security, restrict the access to galera1, which should be
the only machine the script is run on, and do not allow free shell
access, but rather limit execution of commands to what is called by
backup_replica_sync.sh.

Fixes: b525360394ca ("Import SSH keys for Galera backup initialization")
Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>