Use zypper variables to declare mirrors
To avoid repetitive declaration of the download mirrors and to make
switching location specific mirrors easier, install zypper variables
containing the internal and external mirrors responsible for the
location of a given machine and reference those in the repository URLs.
This additionally allows for central overriding of a mirror when using
the command line, useful to circumvent a mirror outage without editing
the .repo files.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>