syslog-ng
Preliminary introduction of syslog-ng as a preparation to replace
rsyslog:

- syslog-ng profile containing initial client and server configuration
  * client configuration utilizes mostly the packaged defaults, and only
    adds routing to the remote server
  * server configuration handles splitting of messages from remote clients
    into a file structure under /var/log/remote, aiming to allow for easy
    browsing of busy client logs and to later add application specific log
    files
- logger role including the server states assigned to logger-prg as an initial
  new logging machine, preparing it as replacement for the legacy syslog role
  currently used with monitor.i.o.o
- client states will be included in the base role in a later patch once the
  solution was evaluated to be good for global use

This new logging implementation will make us utilize syslog messages
according to RFC 5424 instead of the legacy RFC 3164, allowing us to
make use of structured data to pass information between client and
server. This will be useful to filter and split logs on the remote end
based on tags set on the client, further making for a more structured
central logging structure when feeding application logs in addition to
the regular syslog/systemd-journal logs into the system.

Whilst this might have technically been possible with the current
rsyslog based implementation, I found syslog-ng to be more comfortable
to work with in terms of configuration, error handling, and debugging.
As a bonus, we have an AppArmor profile for syslog-ng as part of the default
apparmor-profiles package.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>