Switch to internal-sftp
The "old" sftp subsystem using a separate binary did no longer work on
Tumbleweed as it is located under /usr/libexec there. Instead of
implementing logic to reference the correct libexecdir depending on the
distribution, switch to the modern, internal, sftp subsystem which does
not require calling an external binary and as a bonus makes configuration
of chrooted ssh environments and ACLs easier.
With this patch, our workaround to use scp with -O (to force legacy scp, which
was needed due to the sftp subsystem calling a non-existent path on Tumbleweed)
will no longer be needed.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>