authorized-exec configurations world readable
Configurations need to be readable by the user the permissions apply
for. As one configuration can contain multiple users, regular user/group
ownership is not ideal. POSIX 1e ACLs would be preferable, but the
management states lack functionality for enforcing only a given set of
ACLs. As the information contained in the configuration files is
publicly accessible anyways, set the files to be world readable.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>