Import ipx-proxy1 HAProxy configuration
HAProxy on this machine currently only serves as a reverse proxy for
status.o.o (if the service is currently located in NUE), discard the
obsolete manual configuration in favor of a simplified, Salt managed,
TCP proxy one. If other services are to be hosted in NUE again in the
future, it shall be refactored to a HTTP proxy.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>