Log Asgard link local drops
Logging for dropped packets coming in to the public interface is
generally disabled, as it would cause lots of noise as is expected from
the public internet. For troubleshooting connectivity issues related to
ICMPv6 however, it will be useful to find out about rejections originating
from non-routed, link local, addresses on the public interface as well.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>