test_nftables: templating support
Support .nft.j2 suffixed files in nftables trees and render them through
Jinja before linting and loading the ruleset.
This is a rudimentary implementation which does not support Salt
specific variables / dunder dictionaries due to the direct invocation of
Jinja for simplicity compared to interfacing with Salt, but it might
already useful for some deduplication of code.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>