Increase conntrack hash table size on gateways
Our gateways currently run with 4GB of memory, which leads the kernel to
calculate a default value of 262144.
Whilst no exhaustion resources in the classic form of performance
issues or a table size growing over the maximum (which would result in
conntrack dropping packets) was observed, often return packets from our
web servers to the reverse proxy servers were found to be marked as
INVALID and to not match the ESTABLISHED rule, leading to firewall
packet denials, breaking stateful behavior.
After inspection of netfilter_conntrack_perf.txt and the VyOS
optimization recommendations, increasing the hash size in the form of
buckets, under the premise of our gateway machines being mainly used for
packet routing and firewalling, with a comparatively minor amount of
userspace activity, was deemed to be a good optimization, albeit not yet
being certain to mitigate the observed state tracking issues.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>