Asgard: add if_s2s set
We tend to additionally filter by source or destinaton interface in
addition to address or network to prevent traffic leakage to undesired
paths.
To facilitate such filtering for remote sites, it is needed to whitelist
all site to site tunnel interfaces, as OSPF can adjust routes to send traffic
through an intermediary site in case the primary path to a given site
is down.
Avoid having to list all tunnel interfaces repetitively with each
relevant rule by adding a set which can then easily be referenced in the
rules, and which can also be updated from a single place if necessary.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>