Manage gai.conf to prefer all IPv6 on gateways
By default, glibc prefers only IPv6 GUA addresses over legacy IPv4,
matching the standard set in RFC 6724 *.
This means if only ULA addressing is used on an interface, hosts will be
resolved to IPv4 instead if an A record is present, even if an AAAA
record is present too, and full IPv6 connectivity would be possible
through the ULA source address on an interface.
In our infrastructure this is not desirable - because we generally
prefer *any* IPv6 over legacy IPv4, we also do not facilitate IPv4
connectivity to remote services if not necessary. Some domains however
might still have an A record for compatibility with legacy sites.
This would cause issues on gateway machines which send traffic to remote
sites through their peer to peer tunnel interfaces, which are by design
only equipped with ULA addresses. With glibc resolving to IPv4 instead
of utilizing the ULA address for IPv6 connectivity, such connections
would fail unless IPv6 was forced in all relevant client utilities.
Resolve this by installing a custom /etc/gai.conf which disables the
degredation of ULA addresses, making them get treated equally to GUA
addresses.

* draft-buraglio-6man-rfc6724-update-03 proposes to change this.

On non-gateway machines, ensure the file is absent to be sure the
default configuration is active and no manual overrides alter the
expected behavior.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>