test_nftables: preserve trailing new lines
Jinja will strip trailing new lines, causing nftables files to not
load, and the test to (correctly) fail. In Salt the file seems to be
rendered without altering the trailing new line, even though Salt
does not set this setting in its Jinja environment by default. This
I cannot explain. An alternative would be to add an explicit empty
line to every file in the tree.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>