Asgard: first nftables refactoring
First effort to make use of Jinja templating to deduplicate and
reduce code in our nftables tree.
There is still more to do, and some of the functionality should be
moved to macros, but the aim is to do smaller batches as rendered output
is formatted slightly differently, requiring careful review upon
application.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>