Jumbo frames on Arches
The openSUSE-arches-nfs network makes use of jumbo frames for NFS across
the whole path.
Adjust the Arches networking to facilitate this, but also to allow for
jumbo frames on the physical/bond interfaces by default, and overwrite
logical interfaces used for routed connections to the original default of 1500.
This allows us to later easily add VM networks with jumbo frames if needed.

For better clarity, also explicitly set the high MTU on logical
interfaces which would otherwise inherit it from the parent.

The os-a-nfs interface is capped at 9000, as it is it the maximum on the
NetApp end. Others can make use of the switch maximum at 9216.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>