Initialize Tyr
Internal reverse proxy as well as DNS and NTP forwarder pair for SLC1.
For the beginning, this will forward DNS to hel in PRG2 until a local
recursion setup is available. The HAProxy configuration is prepared for
being populated with more action once needed. Clients in SLC1 will be
adjusted to utilize this as their DNS and NTP servers in a later patch.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>