Correct reverse_pointer() behavior
The reverse_pointer functionality built into the ipaddress library
yields undesired leading zeroes if an IPv6 network is passed.

As this function is currently primarily utilized in the generation of our
configuration for DNS64, it causes reverse lookups to fail, as addresses
cannot be determined to belong to the given reverse domain.
Prior to implementing this functionality the issue was not present, as
the values were manually taken from `ipcalc`, which behaves as expected.

Good:
In  2a07:de40:b27e:64::/96
Out 0.0.0.0.0.0.0.0.4.6.0.0.e.7.2.b.0.4.e.d.7.0.a.2.ip6.arpa.

Bad:
In  2a07:de40:b27e:64::/96
Out 0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.0.4.6.0.0.e.7.2.b.0.4.e.d.7.0.a.2.ip6.arpa.

Correct the logic by replacing it with a contraption found in the related
upstream issue (thank you, @pemensik !):
https://github.com/python/cpython/issues/125641

Fixes: 94220a9047f8 ("Modularize PowerDNS recursor NAT64 configuration")
Fixes: a597f392a6e9 ("Add reverse_pointer function")
Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>