Exclude remote site queries from DNS64
In 200c99242f38 local recursors in SLC1 were introduced, but with an
initial configuration to forward queries to PRG2. In order for DNS64 to
work correctly with this setup, the recursors in PRG2 must not answer
with spoofed AAAA records themselves, as otherwise the already populated
AAAA records, using the PRG2 NAT64 prefix, which is naturally not routable
to the internet in SLC1, would arrive there. Rather have the recursors
skip DNS64 for queries from sites other than the one they are hosted in
completely, in order for the recursors in the other sites to receive the
original record information, and subsequently to perform AAAA spoofing
using the correct site specific NAT64 prefix locally themselves.
An alternative solution would have been to teach all recursors the site
specific NAT64 prefixes of all locations, and to choose the right one
based on the source address, but this was deemed more complex with not
much gained benefit, given the setup to forward queries to another site
not being designed as a permanent solution anyways.

Whilst at it, reformat the Lua script for better diff- and readability,
and improve the explanatory comments.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>