Simplify DNS forwarding setup
This refactors the setup of our "edge" DNS servers. These have
direct connectivity to the internet and run two services:
- an authoritative PowerDNS, serving our own DNS zones
  (listening on port 53)
- a recursive PowerDNS, forwarding queries for names in our own zones
  to the local authoritative servers, and recursing the rest to root
  servers on the internet
  (listening on port 1053)
These type of machines all share the same configuration, with the only
difference being the listening and forwarding addresses.
Adding more of these machines would come with lots of duplication in the
pillar to list the forward zones for every pair just to differentiate
the forwarding addresses. A pillar macro would be a solution, but as we do
not utilize the forwarding configuration for any other purposes, an easier
and potentially less complex solution was found to be incorporating the
concept of our machine type into the state logic, requiring only two
pillar settings per machine, in addition to the list of domains in the
general role pillar.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>