Support IPv4 networks other than .0
So far legacy network IDs were always assumed to be 0, which makes sense
if all networks are /24, but not if smaller chunks are allocated. As
this becomes practice with SLC1, adjust the schema and calling logic to
allow for either specification of only the VIP address (which implicitly
generates the individual gateway IP addresses) in networks.yaml, or now
alternatively the explicit individual addresses for smaller networks.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>