Native IPv4 for mirrorcache-us-db
In our default setup, we "fix" lack of IPv4 through NAT64+DNS64 for
internal hosts which need to connect to such hosts on the internet, but
the MirrorCache service needs to properly detect if a given miror host
on the internet supports IPv6 or only IPv4.
Note mirrorcache-backstage, which does the mirror scanning, runs on
mirrorcache-us-db.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>