Improve drop_user.sh logging
Make the terminal output less cluttered and format it in a way pasteable
into ticket comments as-is to ease the user revocation process.
Log more detailed output to a file, both for easy debugging in case of
issues and for auditing and referencing of past invocations.
The lines written to the new log file are still a bit inconsistent and
its content could potentially be truncated after a while, but I deem it
sufficient for now.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>