Load nftables networks from pillar
States cannot access files in the pillar root, load network data into
the pillar and implement logic to simulate the needed parts of the
pillar in the nftables test to avoid the need for loading a minion.
This was missed in the original templating implementation and only found
upon deployment through the production master.

Fixes: ab109e187832 ("Asgard: first nftables refactoring")
Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>