Drop virt_cluster grain
After previous attempts of dropping this, it should now really no
longer be needed - handling of the "cluster" pillar was implemented in
the formula as a replacement for the "virt_cluster" grain already
several versions ago.
The usage in the network pillar is replaced with a solution deemed
sufficient of a workaround given the explanation in 6f16378e1383.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>