Asgard: refactor GitHub ACLs
This is the first legacy internet ACL refactoring. The goal is to
whitelist outbound access already at the IPv6 stage, aligning with
whitelisting for native IPv6 destinations, and reducing the need for
issuing static IPv4 addresses in TAYGA.
Additionally, the rules shared between VLANs to enable the new ACL are
moved to a chain, which can be included where applicable to reduce
redundant code.
The other acl4_internet_* sets shall later be refactored to the same
structure.

https://progress.opensuse.org/issues/152863

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>