syslog-ng: filter out HAProxy handshake failures
HAProxy tends to log these genericly, upon reading clients connecting with
unsupported cipher suites or applications not completing a handshake are
a suspected cause. As this happens a lot (several 100k lines per day) and
there neither seems to be anything to act upon nor is there any native
functionality to suppress these messages (other than lowering the log
level, which would also suppress messages we are interested in), drop them at
the syslog server, but purposefully match the end of line to preserve
the handshake failure messages containing a specific error (which occur
less frequently and can be considered useful).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>