Introduce powerdns-api-proxy
As we add more services making use of the PowerDNS API, more fine
grained access control is needed to reduce the surface of atatck
presented to a misbehaving or infiltrated API client.
PowerDNS only allows to configure a single, "root", API token and
recommends to use a separate layer for RBAC.
Hence introduce a simple daemon running on the same machine which API
clients will then be configured to connect to using service specific,
restricted, tokens.
This patch only introduces the basic setup, later consumers will be
adjusted to make use of it, with the goal of eventually making the
PowerDNS webserver only listen on a local socket with a root token
not known to any other machines.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>