Block expensive Wiki requests
This request pattern was found to be issued by suspected crawlers so
many times that significant load is caused, making the wikis slow to
respond to legitimate users.
Refuse such requests and make clients issuing them wait 5 seconds before
receiving a return for good measure.

Requests with a session cookie are exempted (which can easily be
spoofed, but is deemed a good compromise to still allow the
functionality for legitimately logged in users until the malicious ones
find out).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>