Reject pagure01 access to www.o.o/openid
This rule is already in place transiently due to pagure01.i.o.o
causing outage of Ipsilon on id.o.o by making over a millions requests
in less than a day.
Rule is added permanently with this patch as there has not been any
input in https://progress.opensuse.org/issues/181646 so far.
Once a solution is found, this should be reverted.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>