pgbouncer: increase file descriptor limit
With the current allowed connection count, PgBouncer would run out of file
descriptors if under load. It does not raise the soft limit on its own.
Note according to systemd.exec(5) this option should not be used. But it
does give an alternative to raising the limit, and it's common practice
to use it for processes run under systemd.

https://progress.opensuse.org/issues/182522

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>