Set fullconn for Berghain backend
Normally HAProxy calculates this to 10% of the default, but with this
backend it would only set it to 1 (which it usually only does for
backends without real servers) - it might be related to the more
complicated use_backend branching. Set it manually to solve alerts about
exhausted maximum backend connections.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>