Confine certificate deployments
The central certificate station makes for convenient and reliable
certificate deployments, but by its design of being able to connect to
all certificate targets also makes for a large backdoor.
Mitigate this situation and lower the possible attack surface which
comes with SSH by restricting remote execution to only the minimum
required for the certificate deployment operations.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>