salt: refactor roles to accomodate syndic setup
Split configuration into three subroles, making the master role only
contain configuration common to all kinds of Salt masters. This prepares
for introduction of a master of masters which should be free from
configuration and data it does not need locally due to delegation to
syndics.
The git fileserver backend is removed whilst at it, as it was no longer
used anyways.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>