bash: hostusage and user for all users
As we encourage administrators to only enter a root shell when needed,
it makes sense to also show the host usage information in a non-root
shell. To further unify the prompts, add a space to the non-root
separator and show the username for root as well (giving an additional
hint to the administrator that they are working in a sensitive session).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>