Load hosts and cert targets into gateway pillar
Allow for templating of hosts including certificate targets in the
nftables configuration of gateways.
The certificate target logic is kept in the common pillar to avoid
additional, gateway role specific, iterations over the data.
The duplication in test_nftables.pl is unfortunate but no better way was
found with the current implementation which renders the files outside of
Salt.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>