Ensure absence of /etc/.gitignore
This file seems to have been created by
/etc/etckeeper/update-ignore.d/01update-ignore
in the past, but now causes etckeeper itself to fail upon execution of
/etc/etckeeper/commit.d/20store-metadata.
As the file only contains ignores for paths which are either no longer
relevant or which are not expensive to track, solve the issue by
ensuring its absence on all machines.
Testing showed that it is not re-created and stays absent once removed.

Link: https://progress.opensuse.org/issues/165117
Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>