Increase socket receive buffer on gateways
Monitoring shows UDP receive queue lengths to spike with values
up to 216000 bytes in the last months. This exceeds 212992 bytes
(which we have as a default value for default and max) and is
suspected to be the cause of a large amount of dropped UDP packets,
particularly affecting conntrackd. Generously help the situation by
doubling the default receive buffer size and further expanding the
maximum.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>