mailserver: change limits to prevent timeouts at end of DATA
... which cause{s,d} external mailservers to think that a mail wasn't
received, and sending them again, resulting in duplicated mails.

Getting the process limit of smtpd and spampd (nearly) in sync means
that in worst case external servers get a timeout on connect. This is
annoying, but not as bad as a timeout after end of DATA (see above).

The most important changes are those in master.cf and the spampd config.

The main.cf changes "can't hurt" and should make the overall situation
better, but are probably less important.

Fixes https://progress.opensuse.org/issues/189534