Hel: rebalance Galera
Historically an application the Galera cluster was used for had issues
with gaps in automatically incrementing fields when load was balanced
across multiple Galera nodes, and subsequently the HAProxy configuration
was set to only utilize one node at a time.
It is doubted that this issue is still relevant with our current
applications using MySQL, and galera1 is finding high utilization -
especially due to it also being used as a backup replication source. Try
balancing again and consider the special role of galera1 by allowing
higher utilization of galera2 and galera3.
This setup also follows the Galera cluster documentation for balancing
with HAProxy.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>