Refactor nodegroups
The master of masters should have node groups covering all locations.
Adjust to a split between global and site specific groups, and add nodes
previously not covered. Unify the group names to have either clusters or
"capabilities".
Hypervisor clusters are unfortunately not split by site in clusters.yaml
and hence kept as-is outside the conditional.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>