salt.master: enforce successful pillar decryption
If decryption of secrets in the pillar using GPG fails, the armored PGP
message would be transmitted to the minion. This is a security risk for
states using encrypted pillars to set passphrases, as decryption failure
could lead to states setting such passphrases to the public PGP message
instead of the intended secret string.
Future versions of Salt will have this set to "True" by default.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>