Refactor Prometheus targets
Switch away from the static pillar configuration to file based service
discovery. This allows targets to be modified without reloading the
service and allows us to more easily split targets per location based
on the minions available in the mine of the Prometheus server minion
responsible for a given location.
Add minor simplifications whilst at it.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>