Prepare basic traffic shaping
Allow for restriction of the bandwidth available to certain hosts/services
to prevent malicious or misbehaving services from congesting the network.
The script and class logic is kept very basic for the beginning -
later it would be nice to extend this with proper priorities. Similarly,
only the nat64 network on Asgard is enabled as a start - if it works as
expected, we should use this to prioritize between zone networks as
well.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>