Asgard/Avalon: add cluster interfaces to s2s set
We have OSPF configured to include the cluster interfaces of the
firewalls in the routing decision for site to site traffic, as this
allows us to have redundant VPN tunnels independent of the firewall
redundancy. Hence, these interfaces need to be included in the firewall
rules allowing traffic through the S2S tunnels, as otherwise traffic
would be rejected if a packet destined for another site was chosen
to be routed through the second firewall instead of directly through
the tunnel.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>