Asgard/Avalon: disable TCP notrack rule
This was introduced as part of the implementation of SYN proxy
functionality. Unfortunately it caused packets to be dropped as INVALID
instead of being processed by the SYN proxy as intended.
The cause is not yet known. Instead of reverting the patch, disable the
notrack rule to effectively revert the SYN proxy but to leave the
configuration for further debugging.

dd8478f58af7 ("Asgard/Avalon: configure SYN proxy")
Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>