Initialize network and VMs for gitlab-runner{3,4}
The two existing runners are often reaching their capacity.
Add some more to allow for more parallel CI jobs, and place them in SLC
as we have much more spare room there. It is assumed that the additional
compute capacity for job load outweighs the added network latency - we
will need to test if that holds true.
Whilst at it, improve the host metadata and set myself as maintainer.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>