hypervisor.cluster: remove stray resources include
The "suse_ha" state already includes this (through a chain of further
includes), if the minion is the designated controller of the cluster.
Explicitly including it causes the resources state to be rendered
regardless of whether the minion is the DC, leading to an error if
dependent resources, which only get applied on the DC, are not present
on the non-DC minion yet. Further, it causes unnecessary duplication of
configuration tests if the state is applied on all minions of the
cluster.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>