powerdns: remove stale zones from autoprimaries
On machines with both the nameserver.recursor and the
nameserver.secondary roles, the autoprimary feature in PowerDNS creates
new zones in the authoritative server, to which the recursive server
then forwards using the domains.yaml data. This setup leaves behind
zones which are no longer managed, causing both stale data, and
consistency issues in case of records moving to different zones
during restructurings. Whilst this is a general lack in the autoprimary
functionality, we can solve it thanks to the existing data in our
combined setup.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>