Add time condition to MySQL replication alert
The daily mysql-backupscript service disables the replication for the
duration of its backup jobs. This is by design and expected, hence limit
the alert to only fire outside of the expected backup period to avoid
confusion and unnecessary alerting noise.
As mysql-backupscript does not expose metrics on its own, choose a time
based condition aligning with the setup of the timer and backup
duration: currently it is triggered at 0300 and runs for 2-3 hours.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>