OSPF: export routes only from primary routers
227451fc3fdf ("Refactor site VPNs to mesh") improved the site to site
VPN topology, however missed that with two connected routers advertising
the same networks at the same cost, the OSPF protocol might install
routes from the secondary router in a remote site, leading to asymmetry
and subsequent packet loss due to remote hosts using the primary router
as their default gateway. Unfortunately, there is no native filter for
the VRRP status (or related metrics, such as Macvlan interface states or
the presence of a VIP address) in BIRD, and FRR did not seem to offer
such functionality either (though we should do more evaluation of FRR
regardless to potentially benefit from handling OSPF and VRRP within a
single and possibly better integrated solution).
One solution to the problem at hand would be reducing the cost to
the secondary routers, however there is no protocol native way to
achieve this from the remote side, and exchanging shell commands between
remote routers is deemed not secure and prone to breakage in case of
connectivity issues.
The typical solution seems to be equipping Keepalived with a notification
script which either completely shuts down BIRD, or which rewrites its
configuration. Both were not deemed good enough:
- all BIRD instances should always be available, as they not only
  advertise routes affected by VRRP, and because startup/shutdown comes
  with connectivity loss
- notification scripts are inefficient and in case of VRRP flapping could
  be executed in mixed up order and potentially even simultaneously
Instead, use Keepalived's FIFO feature and implement a rudimentary "buffer"
system:
- One script constantly constantly consumes the FIFO pipe while
Keepalived is running and updates entries mapping VRRP instances to their
current status in a local database file - Keepalived does have a built in
way to execute a script consuming the FIFO pipe, but here we execute it
independently under systemd for additional robustness and to avoid the need
for implementing Keepalived specific signal handling.
- A second script, run on a schedule, consumes the database as a whole and
generates a list of networks which the router is currently primary for. This
list is then loaded into BIRD for use as an export filter. Currently the
schedule for this is set to ~10 seconds, which should be fast enough without
overwhelming BIRD with reloads in case of many simultaneous updates
(when Keepalived starts for the first time or when a failover of all
instances happens, all instances should enter and report their their
preliminary state with a few seconds) - if it turns out routing updates
need to happen faster for seamless connectivity between sites, we
should follow up with either decreasing the interval or having the
include service triggered on demand during updates.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>